isolates sql_database fixture creation. available only in tests/sql_database
