+++ tests/utils.py
@@ -13,8 +13,6 @@ from dlt.common.storages import FileStorage
 
 from dlt.pipeline.exceptions import SqlClientNotAvailable
 
-from tests.sql_source import SQLAlchemySourceDB
-
 TEST_STORAGE_ROOT = "_storage"
 
 # get env variable with destinations
@@ -57,7 +55,7 @@ def drop_pipeline() -> Iterator[None]:
         Container()[PipelineContext].deactivate()
 
 
-@pytest.fixture(autouse=True)
+@pytest.fixture(autouse=True, scope="session")
 def test_config_providers() -> Iterator[ConfigProvidersContext]:
     """Creates set of config providers where tomls are loaded from tests/.dlt"""
     config_root = "./pipelines/.dlt"
@@ -91,19 +89,6 @@ def preserve_environ() -> None:
     yield
     environ.clear()
     environ.update(saved_environ)
-
-
-@pytest.fixture(scope='session')
-def sql_source_db():
-    db = SQLAlchemySourceDB()
-    try:
-        db.create_schema()
-        db.create_tables()
-        db.insert_data()
-        yield db
-    finally:
-        db.drop_schema()
-
 
 
 def clean_test_storage(init_normalize: bool = False, init_loader: bool = False, mode: str = "t") -> FileStorage:
